Restructuring of the typed session logic
