+++ pyuavcan/presentation/_typed_session/__init__.py
@@ -4,6 +4,9 @@
 # Author: Pavel Kirienko <pavel.kirienko@zubax.com>
 #
 
-from ._base import TypedSession, TypedSessionFinalizer, OutgoingTransferIDCounter
-from ._pub_sub import MessageTypedSession, Publisher, Subscriber
-from ._rpc import ServiceTypedSession
+from ._base import TypedSession, MessageTypedSession, ServiceTypedSession
+from ._base import TypedSessionFinalizer, OutgoingTransferIDCounter
+
+from ._publisher import Publisher
+
+from ._subscriber import Subscriber
