+++ pyuavcan/presentation/_typed_session/_client.py
@@ -1,0 +1,7 @@
+#
+# Copyright (c) 2019 UAVCAN Development Team
+# This software is distributed under the terms of the MIT License.
+# Author: Pavel Kirienko <pavel.kirienko@zubax.com>
+#
+
+from __future__ import annotations
